Run model grid

+++ DRAGON/run_DRAGON.sh
@@ -2,12 +2,20 @@
 
 export DRAGON_DIR=/home/yago/public/git/DRAGON2-Beta_version
 export WORKING_DIR=$(pwd)
-export MODEL=run_2D_DM
+
+if [[ $# != 1 ]];
+then
+    echo "USAGE: run_DRAGON <model_name>"
+    exit
+fi
 
 cd $DRAGON_DIR
-cp $WORKING_DIR/output/$MODEL* $WORKING_DIR/output/bak/
-./DRAGON $WORKING_DIR/input/$MODEL.xml &> $WORKING_DIR/output/$MODEL\_output.txt
-cp output/$MODEL* $WORKING_DIR/output/
+for MODEL in $@
+do
+    [ -f "$WORKING_DIR/output/$MODEL.fits.gz" ] && cp $WORKING_DIR/output/$MODEL* $WORKING_DIR/output/bak/
+    ./DRAGON $WORKING_DIR/input/$MODEL.xml &> $WORKING_DIR/output/$MODEL\_output.txt
+    [ -f "output/$MODEL.fits.gz" ] && cp output/$MODEL* $WORKING_DIR/output/
+done
 
 : <<'Bye'
                                                          ... Paranoy@ rulz! ;^D
